Refactoring: - Extracted `LongList` interface and used it where it's possible. Renamed original class to `AbstractLongList`. - Renamed `fullChunkCleanup` to `releaseChunk` - Moved `updateMinValidIndex` call to `startWriting` MemoryIndexDiskKeyValueStore
