+++ platform-sdk/swirlds-jasperdb/src/test/java/com/swirlds/merkledb/collections/AbstractLongListTest.java
@@ -20,7 +20,7 @@ import static com.swirlds.common.utility.Units.BYTES_TO_MEBIBYTES;
 import static com.swirlds.common.utility.Units.MEBIBYTES_TO_BYTES;
 import static com.swirlds.merkledb.MerkleDbTestUtils.checkDirectMemoryIsCleanedUpToLessThanBaseUsage;
 import static com.swirlds.merkledb.MerkleDbTestUtils.getDirectMemoryUsedBytes;
-import static com.swirlds.merkledb.collections.LongList.FILE_HEADER_SIZE_V2;
+import static com.swirlds.merkledb.collections.AbstractLongList.FILE_HEADER_SIZE_V2;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -37,19 +37,19 @@ import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.api.io.TempDir;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-abstract class AbstractLongListTest<T extends LongList<?>> {
+abstract class AbstractLongListTest<T extends AbstractLongList<?>> {
 
     private static final int OUT_OF_SAMPLE_INDEX = 13_000_123;
     private static final long REPL_VALUE = 42;
     private static final long DEFAULT_VALUE = 0;
 
-    private static LongList<?> longList;
+    private static AbstractLongList<?> longList;
 
     protected int getSampleSize() {
         return 1_000_000;
     }
 
-    protected LongList<?> createLongList() {
+    protected AbstractLongList<?> createLongList() {
         return new LongListHeap();
     }
 
@@ -73,11 +73,11 @@ abstract class AbstractLongListTest<T extends LongList<?>> {
         final long capacity = longList.capacity();
 
         assertEquals(
-                LongList.DEFAULT_MAX_LONGS_TO_STORE,
+                AbstractLongList.DEFAULT_MAX_LONGS_TO_STORE,
                 longList.capacity(),
                 "Capacity should be default it not given explicitly");
         assertEquals(
-                LongList.DEFAULT_NUM_LONGS_PER_CHUNK,
+                AbstractLongList.DEFAULT_NUM_LONGS_PER_CHUNK,
                 longList.getNumLongsPerChunk(),
                 "Num longs per chunk should be default it not given explicitly");
 
@@ -138,7 +138,7 @@ abstract class AbstractLongListTest<T extends LongList<?>> {
             assertEquals(i, readValue, "Longs don't match for " + i + " got [" + readValue + "] should be [" + i + "]");
         }
         // now try and construct a new LongList reading from the file
-        try (final LongList<?> longList2 = createLongListFromFile(file)) {
+        try (final LongList longList2 = createLongListFromFile(file)) {
             // now check data and other attributes
             assertEquals(longList.capacity(), longList2.capacity(), "Unexpected value for longList2.capacity()");
             assertEquals(longList.size(), longList2.size(), "Unexpected value for longList2.size()");
@@ -191,7 +191,7 @@ abstract class AbstractLongListTest<T extends LongList<?>> {
     void chunkSizeFactoryWorks() {
         final int expectedNum = Math.toIntExact(2 * MEBIBYTES_TO_BYTES / Long.BYTES);
 
-        final LongList<?> subject2mbChunks = createLongListWithChunkSizeInMb(2);
+        final AbstractLongList<?> subject2mbChunks = createLongListWithChunkSizeInMb(2);
 
         checkNumLongsPerChunk(subject2mbChunks, expectedNum);
     }
@@ -244,7 +244,7 @@ abstract class AbstractLongListTest<T extends LongList<?>> {
                 "Stream size should match initial sample size");
     }
 
-    protected void checkNumLongsPerChunk(final LongList<?> subject, final int expected) {
+    protected void checkNumLongsPerChunk(final AbstractLongList<?> subject, final int expected) {
         assertEquals(
                 expected,
                 subject.getNumLongsPerChunk(),
